Refactor tests into different modules
