+++ test-manager/src/tests/test_metadata.rs
@@ -6,8 +6,12 @@ pub struct TestMetadata {
     pub name: &'static str,
     pub command: &'static str,
     pub mullvad_client_version: MullvadClientVersion,
-    pub func:
-        Box<dyn Fn(ServiceClient, Box<dyn std::any::Any + Send>) -> BoxFuture<'static, Result<(), Error>>>,
+    pub func: Box<
+        dyn Fn(
+            ServiceClient,
+            Box<dyn std::any::Any + Send>,
+        ) -> BoxFuture<'static, Result<(), Error>>,
+    >,
     pub priority: Option<i32>,
 }
